check if pod is leader when performing scheduled task

+++ src/main/kotlin/no/nav/nks_ai/core/ConversationDeletionJob.kt
@@ -3,19 +3,19 @@ package no.nav.nks_ai.core
 import dev.starry.ktscheduler.job.Job
 import dev.starry.ktscheduler.scheduler.KtScheduler
 import dev.starry.ktscheduler.triggers.DailyTrigger
+import io.github.oshai.kotlinlogging.KotlinLogging
 import io.ktor.client.HttpClient
-import kotlinx.coroutines.runBlocking
 import no.nav.nks_ai.app.isLeader
 import no.nav.nks_ai.core.conversation.ConversationService
+
+private val logger = KotlinLogging.logger { }
 
 class ConversationDeletionJob(
     private val conversationService: ConversationService,
     private val httpClient: HttpClient,
 ) {
-    fun start() = runBlocking {
-        if (isLeader(httpClient)) {
-            scheduleJob()
-        }
+    fun start() {
+        scheduleJob()
     }
 
     // Deletes old conversations at 03:00 every night.
@@ -26,7 +26,13 @@ class ConversationDeletionJob(
             jobId = "DeleteOldConversations",
             trigger = trigger,
             callback = {
-                conversationService.deleteOldConversations()
+                if (isLeader(httpClient)) {
+                    conversationService.deleteOldConversations()
+                } else {
+                    logger.info {
+                        "This instance is not leader. Scheduled job will not be performed"
+                    }
+                }
             }
         )
